feat: remove unused icons and add new contact and portfolio pages
- Deleted multiple unused icon files from the public directory.
- Created a new contact page with a footer component.
- Updated the portfolio page to use a new container component.
- Added a not found page for 404 errors.
- Updated the layout to include a navigation component.
- Created a navigation component with links to home, portfolio, and contact pages.
- Created a footer component with company information and contact details.
- Updated global styles for navigation link colors.
