+++ src/app/contact/page.tsx
@@ -1,0 +1,7 @@
+'use server';
+
+import Contact from '@/containers/contact';
+
+export default async function ContactPage() {
+  return <Contact />;
+}
