+++ src/app/globals.css
@@ -23,10 +23,10 @@
 @layer components {
   /* Navigation */
   .nav-link {
-    @apply h-16 flex items-center text-gray-500 hover:text-primary-700 transition-colors text-sm font-medium;
+    @apply h-16 flex items-center text-gray-500 hover:text-primary-900 transition-colors text-sm font-medium;
   }
   .nav-link.active {
-    @apply text-primary-700;
+    @apply text-primary-900;
   }
 
   /* Buttons */
