+++ src/app/not-found.tsx
@@ -1,0 +1,13 @@
+import Link from 'next/link';
+
+export default function NotFound() {
+  return (
+    <div className="flex flex-col items-center justify-center min-h-screen">
+      <h2 className="text-4xl font-bold mb-4">404</h2>
+      <p className="text-lg mb-4">아무 것도 찾을 수 없습니다.</p>
+      <Link href="/" className="text-blue-500 hover:underline">
+        홈으로 돌아가기
+      </Link>
+    </div>
+  );
+}
